update confidential mode display
